requirements.txt
babel==2.16.0
blinker==1.9.0
boto3==1.35.81
botocore==1.35.81
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorzero==1.1
cryptography==3.3.2
distro==1.5.0
dnspython==2.0.0
Flask==3.1.0
gpg===1.14.0-unknown
gpiozero==1.6.2
idna==2.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Markdown==3.3.4
MarkupSafe==3.0.2
motioneye==0.43.1b2
numpy==1.19.5
picamera2==0.3.12
pidng==4.0.9
piexif==1.1.3
Pillow==8.1.2
pycurl==7.45.4
Pygments==2.7.1
python-apt==2.2.1
python-dateutil==2.9.0.post0
python-prctl==1.7
PyYAML==5.3.1
requests==2.25.1
requests-toolbelt==0.9.1
RPi.GPIO==0.7.0
s3transfer==0.10.4
simplejpeg==1.6.4
six==1.16.0
spidev==3.5
ssh-import-id==5.10
toml==0.10.1
tornado==6.4.2
urllib3==1.26.5
v4l2-python3==0.3.2
Werkzeug==3.1.3
zipp==3.21.0
